Renaming `ElementConfigurationSectionNew` to `ElementConfigurationSection`.
